refactor: split TouchBarSupport.swift into focused files
- TouchBarSupport.swift: view wiring, installer, coordinator, configuration (~122 lines)
- TouchBarController.swift: controller core, identifiers, state observation (~167 lines)
- TouchBarController+Pickers.swift: player/game picker updates and item factories (~192 lines)
- TouchBarController+Actions.swift: main items, play/stop, tap actions (~150 lines)
